reparametrization trial
take relative and absolute perturbations around parameter values for calculation of Jacobian
option of reparametrized A-values: it scales the huge A-values to scales of reaction rate coefficients by calculating reparametrized A-values, i.e. average reaction rate coefficient at the average temperature over all experiments.
however, this option still has some child diseases and should not be used as for now. Also, testing of this reparametrization showed that little progress was made in the levenberg-marquardt optimization routine in estimating pre-exponential factors. Sensitivity was still a problem.

Tools.updateChemistryInput is changed to writeChemistryInput to emphasized that this method only writes a new chemistry input and that trial parameters need to be updated in other places instead of here
Optimization.setCoefficients() uses the Optimization as an argument instead of the coefficients of this Type.
ParameterEstimationDriver.reactorInputsParser2 is a new reactor input parser in which experiments with variable reactor lenghts can be used. (was created mainly to process Stolle's data)